Divided up product screen into several different components in separate folder to make code more readable and make next step, which is fixing grid issues on ProductScreen easier.
