+++ client/src/components/ProductScreen/ChainOptions.jsx
@@ -1,0 +1,9 @@
+import React, {useState} from 'react'
+
+const ChainOptions = () => {
+  return(
+    
+  )
+}
+
+export default ChainOptions
