+++ client/src/components/ProductScreen/NecklaceModal.jsx
